Simplify directory structure
Removing the src and test directories that just contain 1 file.
